move rules back under security group, update docs

+++ proxmoxtf/resource/cluster/firewall/security_group.go
@@ -13,7 +13,9 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 
-	"github.com/bpg/terraform-provider-proxmox/proxmox/cluster/firewall"
+	clusterfirewall "github.com/bpg/terraform-provider-proxmox/proxmox/cluster/firewall"
+	"github.com/bpg/terraform-provider-proxmox/proxmoxtf/resource/firewall"
+	"github.com/bpg/terraform-provider-proxmox/proxmoxtf/structure"
 )
 
 const (
@@ -37,17 +39,9 @@ func SecurityGroup() *schema.Resource {
 			Optional:    true,
 			Default:     dvSecurityGroupComment,
 		},
-		// firewall.MkRule: {
-		// 	Type:        schema.TypeList,
-		// 	Description: "List of rules",
-		// 	Optional:    true,
-		// 	DefaultFunc: func() (interface{}, error) {
-		// 		return []interface{}{}, nil
-		// 	},
-		// 	ForceNew: true,
-		// 	Elem:     &schema.Resource{Schema: firewall2.RuleSchema()},
-		// },
 	}
+
+	structure.MergeSchema(s, firewall.Rules().Schema)
 
 	return &schema.Resource{
 		Schema:        s,
@@ -58,11 +52,11 @@ func SecurityGroup() *schema.Resource {
 	}
 }
 
-func SecurityGroupCreate(ctx context.Context, api firewall.API, d *schema.ResourceData) diag.Diagnostics {
+func SecurityGroupCreate(ctx context.Context, api clusterfirewall.API, d *schema.ResourceData) diag.Diagnostics {
 	comment := d.Get(mkSecurityGroupComment).(string)
 	name := d.Get(mkSecurityGroupName).(string)
 
-	body := &firewall.GroupCreateRequestBody{
+	body := &clusterfirewall.GroupCreateRequestBody{
 		Comment: &comment,
 		Group:   name,
 	}
@@ -72,30 +66,17 @@ func SecurityGroupCreate(ctx context.Context, api firewall.API, d *schema.Resour
 		return diag.FromErr(err)
 	}
 
-	d.SetId(name)
+	diags := firewall.RulesCreate(ctx, api.SecurityGroup(name), d)
+	if diags.HasError() {
+		return diags
+	}
 
-	// diags := RuleCreate(d, func(body *firewall.RuleCreateRequestBody) error {
-	// 	body.Group = &name
-	// 	e := fw.CreateGroupRule(ctx, name, body)
-	// 	if e != nil {
-	// 		return fmt.Errorf("error creating rule: %w", e)
-	// 	}
-	// 	return nil
-	// })
-	// if diags.HasError() {
-	// 	return diags
-	// }
-	//
-	// // reset rules, we re-read them (with proper positions) from the API
-	// err = d.Set(MkRule, nil)
-	// if err != nil {
-	// 	return diag.FromErr(err)
-	// }
+	d.SetId(name)
 
 	return SecurityGroupRead(ctx, api, d)
 }
 
-func SecurityGroupRead(ctx context.Context, api firewall.API, d *schema.ResourceData) diag.Diagnostics {
+func SecurityGroupRead(ctx context.Context, api clusterfirewall.API, d *schema.ResourceData) diag.Diagnostics {
 	var diags diag.Diagnostics
 
 	name := d.Id()
@@ -115,80 +96,19 @@ func SecurityGroupRead(ctx context.Context, api firewall.API, d *schema.Resource
 		}
 	}
 
-	// rules := d.Get(MkRule).([]interface{})
-	// //nolint:nestif
-	// if len(rules) > 0 {
-	// 	// We have rules in the state, so we need to read them from the API
-	// 	for _, v := range rules {
-	// 		ruleMap := v.(map[string]interface{})
-	// 		pos := ruleMap[MkRulePos].(int)
-	//
-	// 		err = readGroupRule(ctx, fw, name, pos, ruleMap)
-	// 		if err != nil {
-	// 			diags = append(diags, diag.FromErr(err)...)
-	// 		}
-	// 	}
-	// } else {
-	// 	ruleIDs, err := fw.ListGroupRules(ctx, name)
-	// 	if err != nil {
-	// 		if strings.Contains(err.Error(), "no such security group") {
-	// 			d.SetId("")
-	// 			return nil
-	// 		}
-	// 		return diag.FromErr(err)
-	// 	}
-	// 	for _, id := range ruleIDs {
-	// 		ruleMap := map[string]interface{}{}
-	// 		err = readGroupRule(ctx, fw, name, id.Pos, ruleMap)
-	// 		if err != nil {
-	// 			diags = append(diags, diag.FromErr(err)...)
-	// 		} else {
-	// 			rules = append(rules, ruleMap)
-	// 		}
-	// 	}
-	// }
+	if diags.HasError() {
+		return diags
+	}
 
-	// if diags.HasError() {
-	// 	return diags
-	// }
-
-	// err = d.Set(MkRule, rules)
-	// diags = append(diags, diag.FromErr(err)...)
-
-	return diags
+	return firewall.RulesRead(ctx, api.SecurityGroup(name), d)
 }
 
-// func readGroupRule(
-// 	ctx context.Context,
-// 	api firewall.SecurityGroup,
-// 	group string,
-// 	pos int,
-// 	ruleMap map[string]interface{},
-// ) error {
-// 	rule, err := api.GetGroupRule(ctx, group, pos)
-// 	if err != nil {
-// 		if strings.Contains(err.Error(), "no such security group") {
-// 			return nil
-// 		}
-// 		return fmt.Errorf("error reading rule %d for group %s: %w", pos, group, err)
-// 	}
-//
-// 	BaseRuleToMap(&rule.BaseRule, ruleMap)
-//
-// 	// pos in the map should be int!
-// 	ruleMap[MkRulePos] = pos
-// 	ruleMap[MkRuleAction] = rule.Action
-// 	ruleMap[MkRuleType] = rule.Type
-//
-// 	return nil
-// }
-
-func SecurityGroupUpdate(ctx context.Context, api firewall.API, d *schema.ResourceData) diag.Diagnostics {
+func SecurityGroupUpdate(ctx context.Context, api clusterfirewall.API, d *schema.ResourceData) diag.Diagnostics {
 	comment := d.Get(mkSecurityGroupComment).(string)
 	newName := d.Get(mkSecurityGroupName).(string)
 	previousName := d.Id()
 
-	body := &firewall.GroupUpdateRequestBody{
+	body := &clusterfirewall.GroupUpdateRequestBody{
 		Group:   newName,
 		ReName:  &previousName,
 		Comment: &comment,
@@ -199,44 +119,23 @@ func SecurityGroupUpdate(ctx context.Context, api firewall.API, d *schema.Resour
 		return diag.FromErr(err)
 	}
 
-	d.SetId(newName)
+	diags := firewall.RulesUpdate(ctx, api.SecurityGroup(previousName), d)
+	if diags.HasError() {
+		return diags
+	}
 
-	// diags := RuleUpdate(d, func(body *firewall.RuleUpdateRequestBody) error {
-	// 	body.Group = &newName
-	// 	e := fw.UpdateGroupRule(ctx, newName, *body.Pos, body)
-	// 	if e != nil {
-	// 		return fmt.Errorf("error updating rule: %w", e)
-	// 	}
-	// 	return nil
-	// })
-	// if diags.HasError() {
-	// 	return diags
-	// }
+	d.SetId(newName)
 
 	return SecurityGroupRead(ctx, api, d)
 }
 
-func SecurityGroupDelete(ctx context.Context, api firewall.API, d *schema.ResourceData) diag.Diagnostics {
+func SecurityGroupDelete(ctx context.Context, api clusterfirewall.API, d *schema.ResourceData) diag.Diagnostics {
 	group := d.Id()
-	//
-	// rules := d.Get(MkRule).([]interface{})
-	// sort.Slice(rules, func(i, j int) bool {
-	// 	ruleI := rules[i].(map[string]interface{})
-	// 	ruleJ := rules[j].(map[string]interface{})
-	// 	return ruleI[MkRulePos].(int) > ruleJ[MkRulePos].(int)
-	// })
-	//
-	// for _, v := range rules {
-	// 	rule := v.(map[string]interface{})
-	// 	pos := rule[MkRulePos].(int)
-	// 	err := api.DeleteGroupRule(ctx, group, pos)
-	// 	if err != nil {
-	// 		if strings.Contains(err.Error(), "no such security group") {
-	// 			break
-	// 		}
-	// 		return diag.FromErr(err)
-	// 	}
-	// }
+
+	diags := firewall.RulesDelete(ctx, api.SecurityGroup(group), d)
+	if diags.HasError() {
+		return diags
+	}
 
 	err := api.DeleteGroup(ctx, group)
 	if err != nil {
